MeCab Split/Mecab Split/requirements.txt
mecab-python-windows==0.996.0
